services/requirements.txt
Django~=4.2.4
djangorestframework~=3.14.0
celery~=5.3.1
dj_database_url
django-celery-results
psycopg2
pytest
pytest-django
pillow
